Moved make_signature method out to decorators.py
